DBOP_DIN is 16bit wide.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@20102 a1c6a512-1295-4272-9138-f99709370657

+++ firmware/export/as3525.h
@@ -96,7 +96,7 @@
 #define DBOP_CTRL         (*(volatile unsigned long *)(DBOP_BASE + 0x08))
 #define DBOP_STAT         (*(volatile unsigned long *)(DBOP_BASE + 0x0C))
 #define DBOP_DOUT         (*(volatile unsigned short*)(DBOP_BASE + 0x10))
-#define DBOP_DIN          (*(volatile unsigned long *)(DBOP_BASE + 0x14))
+#define DBOP_DIN          (*(volatile unsigned short*)(DBOP_BASE + 0x14))
 
 
 /**
